REF: share dtype methods
